tests: normalize functions in message_utils
There was overlap in functionallity and inconsistent naming.

- Removed getAndConsumeFirstReplyMessage(), was duplicated and its
'expectedEndpoint' feature was buggy, as it would still consume the
1st message.

- getAllReplyMessages() was duplicated by returnReplyMsgs() and it
wasn't clear which one would consume (the former didn't consume).
Solved by making the existing consumeAllReplyMessages() return the
messages instead of void. It's now clear that it consumes and avoids
needing more getters since we can just return the consumed.

- Renamed returnReplyMsg() to consumeReplyMsg() consistency with the
above. It consumes just 1 message.

The goal of this commit is only to remove the inconsistencies and code
duplication. The question of whether we want to consume or not is
tackled here.

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>