tests: Wrap sendMessage() with our own testing::sendMessage()
There's lots of duplicated code that's done after sending a message,
like waiting for replies, which is spread throughout all tests. Not
only duplicated, sometimes it's wrong or flaky, and sometimes we forget
to wait, relying on racy behaviour.

The wrapper will be the place for such common code.

For ease of review, this commit only contains the churn and no
functional change yet.

Signed-off-by: Sergio Martins <sergio.martins@kdab.com>